Logistic Regression Model In Progress

+++ PlanarDataClassificationWithOneHiddenLayer/PlanarDataClassificationWithOneHiddenLayer/DatasetVisualizationView.h
@@ -13,7 +13,7 @@
 // Set the dataset to visualize
 // X: 2 x m matrix (features x examples)
 // Y: 1 x m matrix (labels x examples)
-- (void)setDatasetX:(const std::vector<std::vector<double>>&)X 
+- (void)setDatasetX:(const std::vector<std::vector<float>>&)X 
                   Y:(const std::vector<std::vector<uint8_t>>&)Y;
 
 @end
